product-service/modules/catalog/presentation/api/serializers/category_serializer.py
from rest_framework import serializers
from ....infrastructure.models.category_model import CategoryModel


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model  = CategoryModel
        fields = ['id', 'name', 'slug', 'parent', 'description', 'icon']
