product-service/requirements.txt
Django==5.2.12
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
